Pull data from DIMA and report calibration status
Uses RODBC and dplyr to build data frames containing all LPI and gap intercept data relevant to calculating the standard calibration indicators. Creates a data frame for a specified plot where each "observation" is an observer on the plot and the associated variables are the calculated indicators (numeric values) and whether they're calibrated on each (logical values). Also writes it out as a .csv for use elsewhere.

Has a bunch of hooks for adaptation into a Shiny tool of some sort. Does not currently create figures, but that's on the to-do list.